requirements.txt
django==1.7.4
djangorestframework
django-cors-headers==1.1.0